Simplify color files, make apple theme default on iOS
