+++ packages/theming/apple-theme/src/appleTheme.ios.ts
@@ -1,7 +1,6 @@
 import { Theme, Spacing } from '@fluentui-react-native/theme-types';
-import { paletteFromAppleColors } from './appleTheme.colors.ios';
+import { paletteFromAppleColors } from './appleColors.ios';
 import { appleTypography } from './appleTypography.ios';
-import { getFluentUIAppleDarkPalette, getFluentUIAppleLightPalette } from './fluentAppleColors.ios';
 
 export function appleSpacing(): Spacing {
   return { s2: '4px', s1: '8px', m: '16px', l1: '20px', l2: '32px' };
@@ -43,7 +42,7 @@ export const appleComponents = {
 };
 
 export const BaseAppleLightThemeIOS: Theme = {
-  colors: paletteFromAppleColors(getFluentUIAppleLightPalette()),
+  colors: paletteFromAppleColors(false),
   typography: appleTypography(),
   spacing: appleSpacing(),
   components: appleComponents,
@@ -51,7 +50,7 @@ export const BaseAppleLightThemeIOS: Theme = {
 };
 
 export const BaseAppleDarkThemeIOS: Theme = {
-  colors: paletteFromAppleColors(getFluentUIAppleDarkPalette()),
+  colors: paletteFromAppleColors(true),
   typography: appleTypography(),
   spacing: appleSpacing(),
   components: appleComponents,
